requirements.txt
streamlit
# torch
# torchvision
timm==1.0.3
opencv-python-headless
albumentations==1.3.1
pillow
numpy
matplotlib